Hide top banners in adfree experience
This is a basic implementation; it hides the banner and its space, but there are still questions about the best way to remove the ad space for the best user experience. Here, for example, the page initially renders with the adslot space and then withdraws it, which is suboptimal for the adfree user. We should explore other ways to hide and show these commercial content containers.